State changes and more (#4)
* Minor stack comparison change.

* Changes to state maps, model files, added functionality to crushing table, and more.
Changed state maps to only contain variants for the inventory. Potentially add a separate bimap for state to save to meta, but in a completely separate map.
Added sounds.
Slight changes to implementations of electic nodes.
Removed BlockProperties, moved all properties to package-level near their respective blocks. They should be inside the objects, but that causes an error with order of initialization, unfortunately.
Removed several methods on vectors that duplicate existing methods and are not used anywhere.

* Formatting and merge error fixes